bruh

+++ suckless/dwm/dwm.c
@@ -745,7 +745,7 @@ drawbar(Monitor *m)
 	if ((w = m->ww - tw - x) > bh) {
 		if (m->sel) {
             		/* fix overflow when window name is bigger than window width */
-			int mid = (w - TEXTW(m->sel->name)) / 2;
+			int mid = (w - (int)TEXTW(m->sel->name)) / 2;
 			/* make sure name will not overlap on tags even when it is very long */
 			mid = mid >= lrpad / 2 ? mid : lrpad / 2;
 			drw_setscheme(drw, scheme[m == selmon ? SchemeSel : SchemeNorm]);
